# functions/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "EventManager docs",
    "description": "Documentation for EventManager API.",
    "version": "1.0.0"
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "surname": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "createdEvents": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "subscribedEvents": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": ["name", "surname", "email"]
      },
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": ["message"]
      },
      "ExtendedError": {
        "allOf": [
          { "$ref": "#/components/schemas/Error" },
          {
            "properties": {
              "location": {
                "type": "string"
              },
              "param": {
                "type": "string"
              }
            }
          }
        ]
      },
...
    }
  },
  "paths": {
    "/signup": {
      "post": {
        "tags": ["Auth"],
        "summary": "Subscribe a new user.",
        "description": "Subscribe a new user in the system saving the given informations.",
        "parameters": [
          {
            "name": "email",
            "in": "body",
            "description": "Email of the user.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "password",
            "in": "body",
            "description": "Password of the user.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "password",
              "minLength": 6
            }
          },
          {
            "name": "name",
            "in": "body",
            "description": "Name of the user.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "surname",
            "in": "body",
            "description": "Surname of the user.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Succesful operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Incorrect parameters.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtendedError"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable entity.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/signin": {
      "post": {
        "tags": ["Auth"],
        "summary": "Signin with email and password.",
        "description": "Returns a JWT token to use to access resources.",
        "parameters": [
          {
            "name": "email",
            "in": "body",
            "description": "Email of the user.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "password",
            "in": "body",
            "description": "Password of the user.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "password",
              "minLength": 6
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Succesful operation.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string"
                    },
                    "id": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Incorrect parameters.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtendedError"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable entity.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}